+++ test/helpers/ExpectHelper.js
@@ -1,0 +1,12 @@
+const assertInvalidOpcode = require('../helpers/assertInvalidOpcode');
+
+async function expectThrow(shortTx, call) {
+  try {
+    await call();
+    throw new Error('Did not throw');
+  } catch (e) {
+    assertInvalidOpcode(e);
+  }
+}
+
+module.exports.expectThrow = expectThrow;
